Okay so, I completely changed the races thing
I decided to call them "neural adaptions" and it will give you the same perks as before, but your race is always human as to better fit into the lore.

You will be a half angel human, and the protagonist of the story. You'll have to level up too and fight things.

I'm not sure how in depth this game will go, as I do not even have a clue how to give you the option to save your game

This is all for tonight though!